+++ src/site/src/app/components/docs-content-page/docs-content-page.component.html
@@ -1,4 +1,4 @@
-<div #scrollSpy class="ids-container">
+<div scrollSpy (sectionChange)="onSectionChange($event)" class="ids-container">
 <div class="ids-row--col-sm-10 ids-row--col-xl-7 ids-row--offset-sm-1 page-content page-section">
 
   <div class="ids-banner ids-banner--top-full-width" [hidden]="!showWarning">
